Move input container to it's own file
