+++ packages/text-input/src/TextInput.tsx
@@ -1,18 +1,17 @@
 import { css } from '@emotion/css';
 import { useFocusRing } from '@spark-web/a11y';
-import type { BoxProps } from '@spark-web/box';
 import { Box } from '@spark-web/box';
 import type { FieldState } from '@spark-web/field';
 import { useFieldContext } from '@spark-web/field';
 import { useText } from '@spark-web/text';
 import { useTheme } from '@spark-web/theme';
 import type { DataAttributeMap } from '@spark-web/utils/internal';
-import type { InputHTMLAttributes, ReactElement } from 'react';
+import type { InputHTMLAttributes } from 'react';
 import { forwardRef } from 'react';
 
 import type { AdornmentsAsChildren } from './childrenToAdornments';
 import { childrenToAdornments } from './childrenToAdornments';
-import type { InputAdornmentProps } from './InputAdornment';
+import { InputContainer } from './InputContainer';
 
 type ValidTypes =
   | 'text'
@@ -122,50 +121,4 @@ export const useInput = ({ disabled }: UseInputProps) => {
       },
     },
   } as const;
-};
-
-export type InputContainerProps = {
-  startAdornment?: ReactElement<InputAdornmentProps> | null;
-  endAdornment?: ReactElement<InputAdornmentProps> | null;
-} & Omit<BoxProps, 'background' | 'position'>;
-
-export const InputContainer = ({
-  children,
-  startAdornment,
-  endAdornment,
-  ...boxProps
-}: InputContainerProps) => {
-  const [{ disabled, invalid }] = useFieldContext();
-
-  return (
-    <Box
-      position="relative"
-      background={disabled ? 'inputDisabled' : 'input'}
-      {...boxProps}
-    >
-      {startAdornment}
-      {children}
-      <FocusIndicator invalid={invalid} />
-      {endAdornment}
-    </Box>
-  );
-};
-
-const FocusIndicator = ({ invalid }: { invalid: boolean }) => {
-  return (
-    <Box
-      aria-hidden="true"
-      as="span"
-      data={{ 'focus-indicator': 'true' }}
-      // Styles
-      border={invalid ? 'critical' : 'field'}
-      borderRadius="small"
-      position="absolute"
-      bottom={0}
-      left={0}
-      right={0}
-      top={0}
-      className={css({ pointerEvents: 'none' })}
-    />
-  );
 };
